fix: floats not long for data sources
Better model creation helped catch these issues.

+++ src/main/java/jimlind/filmlinkd/system/letterboxd/model/LBAbstractSearchItem.java
@@ -3,7 +3,7 @@ package jimlind.filmlinkd.system.letterboxd.model;
 // https://api-docs.letterboxd.com/#/schemas/AbstractSearchItem
 public class LBAbstractSearchItem {
   public String type;
-  public long score;
+  public float score;
   public LBContributor contributor;
   public LBFilmSummary film;
   public LBListSummary list;
